Add logic for configuration of model object on initial creation
ModelObjectConfigurer was added. The class could be used to initialize 
some default values on the model object. For example we could set the 
current logged in user if there is a field and/accessor methods for 
that on the model object.

To achieve the above:

```java
@Configuration
public class WebformConfigurerImpl implements WebformConfigurer{
    private static class PostPreconfigurer implements ModelObjectConfigurer<Post>{
        @Override
        public Post configure(Post post) {
            // Configure the Post here
            return post;
        }
    }

    @Override
    public void addModelObjectConfigurers(ModelObjectConfigurerService service) {
        service.addConfigurer(Post.class, new PostPreconfigurer());
    }
}
```